Deleting and archiving boards

+++ Delta3Mini/api.py
@@ -29,11 +29,17 @@ def delete():
     json_data = request.json
     board_to_delete = Board.query.get_or_404(json_data['id'])
 
-    try:
-        db.session.delete(board_to_delete)
-        db.session.commit()
-        return jsonify({'result': 'success'})
-    except:
+    user = User.query.filter_by(username=json_data['username']).first()
+
+    if board_to_delete in user.boards:
+
+        try:
+            db.session.delete(board_to_delete)
+            db.session.commit()
+            return jsonify({'result': 'success'})
+        except:
+            return 'There was a problem deleting that task'
+    else:
         return 'There was a problem deleting that task'
 
 
@@ -87,6 +93,26 @@ def generateList():
     return jsonify({'result': status})
 
 
+@apibp.route('/api/archive', methods=["POST"])
+def archive():
+    json_data = request.json
+    board_to_archive = Board.query.get_or_404(json_data['id'])
+
+    user = User.query.filter_by(username=json_data['username']).first()
+
+    if board_to_archive in user.boards:
+        board_to_archive.archived = True
+        try:
+            db.session.commit()
+            db.session.close()
+            return jsonify({'result': 'success'})
+        except:
+            return 'There was a problem deleting that task'
+
+    else:
+        return 'There was a problem deleting that task'
+
+
 @apibp.route('/api/generateCard', methods=["POST"])
 def generateCard():
     json_data = request.json
@@ -102,3 +128,22 @@ def generateCard():
         status = 'this card couldn\'t have been added'
     db.session.close()
     return jsonify({'result': status})
+
+@apibp.route('/api/getBoardInfo', methods=["POST"])
+def getBoardInfo():
+    json_data = request.json
+    board = Board.query.filter_by(id=json_data['board_id']).first()
+    return jsonify(board=board.serialize)
+
+@apibp.route('/api/unarchiveBoard', methods=["POST"])
+def unarchiveBoard():
+    json_data = request.json
+    board = Board.query.filter_by(id=json_data['board_id']).first()
+    board.archived = False
+
+    try:
+        db.session.commit()
+        db.session.close()
+        return jsonify({'result': 'success'})
+    except:
+        return 'There was a problem deleting that task'
